Switching to using Splinter as a dependency.
Using the module `Splinter` as a dependency. This allows for borrowing a lot of
useful methods already available by this module, and let's us drop the
`utils.py` helper methods.

Also, every `Verb` is now a method and not a class, which allows us to return
web elements for some of the verbs, where it makes sense to return something.